Change climbing logic to single command
