--- Robot2019/src/main/java/frc/robot/commands/RaiseClimber.java
@@ -1,53 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj.command.Command;
-import frc.robot.subsystems.Climber;
-
-/**
- * An example command.  You can replace me with your own command.
- */
-public class RaiseClimber extends Command {
-  private Climber climber;
-  public RaiseClimber(Climber climber) {
-    // Use requires() here to declare subsystem dependencies
-    requires(climber);
-    this.climber = climber;
-  }
-
-  // Called just before this Command runs the first time
-  @Override
-  protected void initialize() {
-  }
-
-  // Called repeatedly when this Command is scheduled to run
-  @Override
-  protected void execute() {
-    climber.runClimber(0.5);    // We want to raise the climber slowly.
-  }
-
-  // Make this return true when this Command no longer needs to run execute()
-  @Override
-  protected boolean isFinished() {
-    return !climber.needToClimb();
-  }
-
-  // Called once after isFinished returns true
-  @Override
-  protected void end() {
-    climber.stopClimber();
-  }
-
-  // Called when another command which requires one or more of the same
-  // subsystems is scheduled to run
-  @Override
-  protected void interrupted() {
-    end();
-  }
-}
